fix: Solve error load component.
