+++ src/views/Home/index.tsx
@@ -1,52 +1,23 @@
 'use client'
 
-import { useMemo } from 'react'
-import { useDebouncedCallback } from 'use-debounce'
-import { usePathname, useRouter, useSearchParams } from 'next/navigation'
+import { Suspense } from 'react'
 
 import { useGetUsers } from '@/hooks/useGetUsers'
+import Search from './components/search'
 
-const Home = () => {
-  const router = useRouter()
-  const pathname = usePathname()
-  const searchParams = useSearchParams()
+interface HomeProps {
+  search: string
+}
 
-  const params = useMemo(() => {
-    const search = searchParams.get('search')
-
-    return {
-      name: search || ''
-    }
-  }, [searchParams])
-
-  const { data: users, isLoading, isError } = useGetUsers(params)
-
-  const handleSearch = useDebouncedCallback((search: string) => {
-    const params = new URLSearchParams(searchParams)
-
-    if (search) {
-      params.set('search', search)
-    } else {
-      params.delete('search')
-    }
-
-    router.replace(`${pathname}?${params.toString()}`)
-  }, 300)
+const Home = (props: HomeProps) => {
+  const { data: users, isLoading, isError } = useGetUsers({ name: props.search })
 
   return (
     <div className='flex items-center justify-center h-screen'>
       <div>
-        <div className='flex flex-col gap-1'>
-          <label>Usuários</label>
-          <input
-            name='search'
-            className='border-2 border-gray-300 rounded-md p-2 min-w-[250px]'
-            type='text'
-            placeholder='Buscar usuário'
-            defaultValue={params.name}
-            onChange={e => handleSearch(e.target.value)}
-          />
-        </div>
+        <Suspense fallback={<div>Carregando...</div>}>
+          <Search />
+        </Suspense>
         <div>
           {isLoading ? (
             <div className='border-2 border-gray-300 rounded-md p-2 mt-2'>
